% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBuildR.R
\name{queryBuildR}
\alias{queryBuildR}
\title{Create a Query Builder widget}
\usage{
queryBuildR(filters, rules = NULL, plugins = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{filters}{List specifying the variables, conditions and value ranges for building the queries.}

\item{rules}{List initializing the set of rules for the Query Builder widget.}
}
\description{
Create a Query Builder widget
}
\examples{
#### Iris data example.
# Load data
data(iris)
filters<-getFiltersFromTable(iris)
# Create Query Builder widget
queryBuildR(filters)

queryBuildR(getFiltersFromTable(iris,
                                column_opts = list(Species=list(type='select',
                                                                plugin='selectize',
                                operators=list('equal','not_equal','contains', 'in', 'not_in')))
                                )
           )

}
