% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBuildR.R
\name{getFiltersFromTable}
\alias{getFiltersFromTable}
\title{Returns a template for the Query Builder filter on the basis of a Data Frame}
\usage{
getFiltersFromTable(data)
}
\arguments{
\item{data}{Data Frame from which the list of filters will be created.}
}
\description{
Returns a template for the Query Builder filter on the basis of a Data Frame
}
\examples{
#### Iris data example.
# Load data
data(iris)
filters<-getFiltersFromTable(iris)
# Set initial rules to NULL
rules<-NULL
# Create Query Builder widget
queryBuildR(rules,filters)
}
