% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBuildR.R
\name{queryBuildR}
\alias{queryBuildR}
\title{Create a Query Builder widget}
\usage{
queryBuildR(rules, filters, width = NULL, height = NULL)
}
\arguments{
\item{rules}{List initializing the set of rules for the Query Builder widget.}

\item{filters}{List specifying the variables, conditions and value ranges for building the queries.}
}
\description{
Create a Query Builder widget
}
\examples{
#### Iris data example.
# Load data
data(iris)
# Rename columns and get filters
colnames(iris)<-c("sepal_length","sepal_width","petal_length","petal_width","species")
filters<-getFiltersFromTable(iris)
# Set initial rules to NULL
rules<-NULL
# Create Query Builder widget
queryBuildR(rules,filters)

}
